[ISSUE-1254] Move PublicAsset APIs to jmap-extensions-api
