+++ tmail-backend/jmap/extensions-api/pom.xml
@@ -29,6 +29,11 @@
             <artifactId>james-server-jmap-rfc-8621</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.github.f4b6a3</groupId>
+            <artifactId>uuid-creator</artifactId>
+            <version>${uuid-creator.version}</version>
+        </dependency>
+        <dependency>
             <groupId>${james.groupId}</groupId>
             <artifactId>testing-base</artifactId>
             <scope>test</scope>
