+++ tmail-backend/jmap/extensions-cassandra/src/test/java/com/linagora/tmail/james/jmap/publicAsset/CassandraPublicAssetServiceTest.java
@@ -5,10 +5,11 @@ import org.apache.james.backends.cassandra.CassandraClusterExtension;
 import org.apache.james.blob.api.BucketName;
 import org.apache.james.blob.memory.MemoryBlobStoreDAO;
 import org.apache.james.server.blob.deduplication.DeDuplicationBlobStore;
+import org.apache.james.util.Size;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import com.linagora.tmail.james.jmap.JMAPExtensionConfiguration;
+import com.linagora.tmail.james.jmap.PublicAssetTotalSizeLimit;
 
 class CassandraPublicAssetServiceTest implements PublicAssetServiceContract {
 
@@ -21,14 +22,15 @@ class CassandraPublicAssetServiceTest implements PublicAssetServiceContract {
 
     @BeforeEach
     void setup(CassandraCluster cassandra) {
-        JMAPExtensionConfiguration jmapExtensionConfiguration = new JMAPExtensionConfiguration(JMAPExtensionConfiguration.PUBLIC_ASSET_TOTAL_SIZE_LIMIT_DEFAULT());
+        PublicAssetTotalSizeLimit publicAssetTotalSizeLimit =
+            PublicAssetTotalSizeLimit.of(Size.of(20L, Size.Unit.M)).get();
         publicAssetRepository = new CassandraPublicAssetRepository(
             new CassandraPublicAssetDAO(cassandra.getConf(), blobIdFactory()),
             new DeDuplicationBlobStore(new MemoryBlobStoreDAO(), BucketName.DEFAULT, blobIdFactory()),
-            jmapExtensionConfiguration,
+            publicAssetTotalSizeLimit,
             PublicAssetRepositoryContract.PUBLIC_ASSET_URI_PREFIX());
 
-        publicAssetSetService = new PublicAssetSetService(PublicAssetServiceContract.identityRepository(), publicAssetRepository, jmapExtensionConfiguration);
+        publicAssetSetService = new PublicAssetSetService(PublicAssetServiceContract.identityRepository(), publicAssetRepository, publicAssetTotalSizeLimit);
     }
 
     @Override
