+++ tmail-backend/jmap/extensions/src/main/scala/com/linagora/tmail/james/jmap/JMAPExtensionConfiguration.scala
@@ -9,10 +9,8 @@ import com.google.common.base.Preconditions
 import com.linagora.tmail.james.jmap.JMAPExtensionConfiguration.{CALENDAR_EVENT_REPLY_SUPPORTED_LANGUAGES_DEFAULT, PUBLIC_ASSET_TOTAL_SIZE_LIMIT_DEFAULT, TICKET_IP_VALIDATION_ENABLED}
 import com.linagora.tmail.james.jmap.method.CalendarEventReplySupportedLanguage.LANGUAGE_DEFAULT
 import com.linagora.tmail.james.jmap.model.LanguageLocation
-import eu.timepit.refined
 import org.apache.commons.configuration2.Configuration
 import org.apache.james.core.MailAddress
-import org.apache.james.jmap.core.UnsignedInt.{UnsignedInt, UnsignedIntConstraint}
 import org.apache.james.server.core.MissingArgumentException
 import org.apache.james.util.{DurationParser, Size}
 
@@ -72,13 +70,6 @@ object JMAPExtensionConfiguration {
   }
 }
 
-object PublicAssetTotalSizeLimit {
-  def of(size: Size): Try[PublicAssetTotalSizeLimit] = refined.refineV[UnsignedIntConstraint](size.asBytes()) match {
-    case Right(value) => Success(PublicAssetTotalSizeLimit(value))
-    case Left(error) => Failure(new NumberFormatException(error))
-  }
-}
-
 case class JMAPExtensionConfiguration(publicAssetTotalSizeLimit: PublicAssetTotalSizeLimit = PUBLIC_ASSET_TOTAL_SIZE_LIMIT_DEFAULT,
                                       supportMailAddress: Option[MailAddress] = None,
                                       supportHttpLink: Option[URL] = None,
@@ -89,10 +80,6 @@ case class JMAPExtensionConfiguration(publicAssetTotalSizeLimit: PublicAssetTota
   def this(publicAssetTotalSizeLimit: PublicAssetTotalSizeLimit) = {
     this(publicAssetTotalSizeLimit, Option.empty)
   }
-}
-
-case class PublicAssetTotalSizeLimit(value: UnsignedInt) {
-  def asLong(): Long = value.value
 }
 
 case class TicketIpValidationEnable(value: Boolean)
