+++ tmail-backend/jmap/extensions/src/test/scala/com/linagora/tmail/james/jmap/publicAsset/PublicAssetDeletionTaskStepTest.scala
@@ -26,7 +26,7 @@ class PublicAssetDeletionTaskStepTest {
     publicAssetRepository = new MemoryPublicAssetRepository(new DeDuplicationBlobStore(new MemoryBlobStoreDAO,
         BucketName.DEFAULT,
         new PlainBlobId.Factory()),
-      JMAPExtensionConfiguration(),
+      JMAPExtensionConfiguration.PUBLIC_ASSET_TOTAL_SIZE_LIMIT_DEFAULT,
       PUBLIC_ASSET_URI_PREFIX)
     publicAssetDeletionTaskStep = new PublicAssetDeletionTaskStep(publicAssetRepository);
   }
